refactor: use result type more widely
